requirements.txt
DrissionPage>=4.0.0
loguru>=0.7.0
requests>=2.31.0
tabulate>=0.9.0
pyyaml>=6.0
rich>=13.7.0